requirements.txt
apache-airflow
apache-airflow-providers-postgres
pandas
bs4 
requests
selenium
psycopg2-binary
sqlalchemy
